+++ deployer/__main__.py
@@ -1,6 +1,6 @@
 # Import the various subcommands here, they will be automatically
 # registered into the app
-import deployer.central_grafana  # noqa: F401
+import deployer.grafana.central_grafana  # noqa: F401
 import deployer.debug  # noqa: F401
 import deployer.deployer  # noqa: F401
 import deployer.generate.generate_aws_cluster  # noqa: F401
